p29.32 refactor: thread ids.UserID through the audit actor
Finish the typed-ID capstone: the change-actor id was the last deferred
bare-int64 entity id. Convert it everywhere it is the CHANGE actor:

- sqlc: Change.ActorID and InsertChangeParams.ActorID -> ids.UserID;
  the two version-history rows (SplitVersionHistoryRow,
  TransactionVersionHistoryRow) that surface c.actor_id -> ids.UserID.
  database/sql binds a named ~int64 to SQLite fine; no cast needed.
- store: drop the int64(actor.ID) cast at the InsertChange bind now that
  Actor.ID (ids.UserID since p29.24) and the param share a type.
- store/history: HistoryEntry.ActorID and the ensure() closure param
  -> ids.UserID.

Purely types; zero behavior change. models.go now has no Actor.*int64
and no non-PK bare-int64 id (every remaining ID int64 is a *_versions.ID
snapshot PK, which rule 15 keeps int64).

+++ internal/store/store.go
@@ -137,7 +137,7 @@ func (s *Store) write(
 	q := s.q.WithTx(tx)
 
 	changeID, err := q.InsertChange(ctx, sqlc.InsertChangeParams{
-		ActorID: int64(actor.ID),
+		ActorID: actor.ID,
 		At:      s.now().Format(time.RFC3339Nano),
 		Kind:    kind,
 		Note:    nullString(note),
